feat: split SalaryPage.tsx into focused subcomponents
Break the monolithic SalaryPage.tsx (~1840 lines) into focused,
independently testable subcomponents under web/src/pages/salary/:
MonthView, YearView, ConfigEditor, TrekktabellEditor, AssignmentsList
and VacationCard. All shared fetches and mutations are lifted into a
single useSalaryData hook (one source of truth, no duplicated fetch
logic), and SalaryPage is reduced to a thin tab router owning only the
active tab and selected month/year state.

Month/year navigation is now handled by a single changeMonth() helper
that updates selectedMonth and selectedYear together, so the two stay
in sync across year boundaries (the previous desync risk is removed).

Refactor only: no API, schema, or user-visible behavior changes; all
strings still go through react-i18next in the existing salary namespace.

Bead: Hytte-hsa53
